requirements.txt
pandas>=2.0
openpyxl>=3.1
pdfplumber>=0.11
# Opcionales:
tabula-py>=2.9
pdf2image>=1.17
pytesseract>=0.3
Pillow>=10.0
